myproject/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tinymce==3.6.1
django-wysiwyg-redactor==0.5.1
djangorestframework==3.14.0
idna==3.4
jsmin==3.0.1
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.4
